Reorganized model.go
