+++ src/github.com/cjlucas/unnamedcast/server/db/query.go
@@ -1,0 +1,12 @@
+package db
+
+import "gopkg.in/mgo.v2/bson"
+
+type Query struct {
+	Filter         bson.M
+	SortField      string
+	SortDesc       bool
+	SelectedFields []string
+	OmittedFields  []string
+	Limit          int
+}
